Update rubocop and rubocop-ast and widen the dependency version range.
It doesn't need to pin a version because the dependency in code_keeper is on private apis.
A private api doesn't gurantee a backward compatibility, it may be
unnecessary to pin even a major version.

Another reason to keep step with rubocop and rubocop-ast updates is to
catch up Ruby updates. The newest ruby isn't supported by them.